+++ graphics/GraphicsClient/GraphicsClient.vcxproj
@@ -22,7 +22,7 @@
     <VCProjectVersion>15.0</VCProjectVersion>
     <ProjectGuid>{A9924152-8D03-4207-BE62-CE1F5A6ACBE9}</ProjectGuid>
     <RootNamespace>GraphicsClient</RootNamespace>
-    <WindowsTargetPlatformVersion>10.0.17763.0</WindowsTargetPlatformVersion>
+    <WindowsTargetPlatformVersion>10.0.16299.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -150,6 +150,7 @@
     <ClInclude Include="..\..\network\StringParser.h" />
     <ClInclude Include="..\..\network\Tile.h" />
     <ClInclude Include="..\..\network\Walls.h" />
+    <ClInclude Include="AbstractGraphicsEngine.h" />
     <ClInclude Include="Animation.h" />
     <ClInclude Include="AnimationChannel.h" />
     <ClInclude Include="AnimationPlayer.h" />
@@ -161,6 +162,7 @@
     <ClInclude Include="FogGenerator.h" />
     <ClInclude Include="Geometry.h" />
     <ClInclude Include="ItemModelType.h" />
+    <ClInclude Include="InGameGraphicsEngine.h" />
     <ClInclude Include="Keyframe.h" />
     <ClInclude Include="MeshEntry.h" />
     <ClInclude Include="Node.h" />
@@ -188,6 +190,7 @@
     <ClCompile Include="..\..\network\StringParser.cpp" />
     <ClCompile Include="..\..\network\Tile.cpp" />
     <ClCompile Include="..\..\network\Walls.cpp" />
+    <ClCompile Include="AbstractGraphicsEngine.cpp" />
     <ClCompile Include="Animation.cpp" />
     <ClCompile Include="AnimationChannel.cpp" />
     <ClCompile Include="AnimationPlayer.cpp" />
@@ -197,6 +200,7 @@
     <ClCompile Include="Bone.cpp" />
     <ClCompile Include="FogGenerator.cpp" />
     <ClCompile Include="Geometry.cpp" />
+    <ClCompile Include="InGameGraphicsEngine.cpp" />
     <ClCompile Include="Keyframe.cpp" />
     <ClCompile Include="MeshEntry.cpp" />
     <ClCompile Include="Node.cpp" />
