+++ graphics/GraphicsClient/GraphicsClient.vcxproj.filters
@@ -103,6 +103,10 @@
     <ClCompile Include="..\..\network\Walls.cpp">
       <Filter>Source Files\Networking and State Tracking</Filter>
     </ClCompile>
+    <ClCompile Include="AbstractGraphicsEngine.cpp">
+      <Filter>Source Files\Client Graphics Rendering</Filter>
+    </ClCompile>
+    <ClCompile Include="InGameGraphicsEngine.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Animation.h">
@@ -237,6 +241,10 @@
     <ClInclude Include="ItemModelType.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="AbstractGraphicsEngine.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="InGameGraphicsEngine.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config">
